refactor: clean up util packages
