+++ .config/.mockery.yaml
@@ -17,9 +17,9 @@ packages:
   github.com/ruslanSorokin/lock-manager/internal/lock-manager/metric:
     interfaces:
       ServiceMetricI:
-  github.com/ruslanSorokin/lock-manager/internal/pkg/apputil:
+  github.com/ruslanSorokin/lock-manager/internal/pkg/util/app:
     interfaces:
       MetricI:
-  github.com/ruslanSorokin/lock-manager/internal/pkg/grpcutil:
+  github.com/ruslanSorokin/lock-manager/internal/pkg/util/grpc:
     interfaces:
       RecoveryMetricI:
