+++ cmd/lock-manager/config.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/app"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/apputil"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/grpcutil"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/promutil"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/redisconn"
+	redisconn "github.com/ruslanSorokin/lock-manager/internal/pkg/conn/redis"
+	apputil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/app"
+	grpcutil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/grpc"
+	promutil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/prom"
 )
 
 type Config struct {
