+++ internal/lock-manager/app/app.go
@@ -12,9 +12,9 @@ import (
 
 	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/handler/igrpc"
 	iservice "github.com/ruslanSorokin/lock-manager/internal/lock-manager/service"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/apputil"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/promutil"
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/redisconn"
+	redisconn "github.com/ruslanSorokin/lock-manager/internal/pkg/conn/redis"
+	apputil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/app"
+	promutil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/prom"
 )
 
 type (
