+++ internal/pkg/conn/redis/error.go
