+++ internal/pkg/util/grpc/iprom/metric.go
@@ -3,8 +3,7 @@ package iprom
 import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
-
-	"github.com/ruslanSorokin/lock-manager/internal/pkg/grpcutil"
+	grpcutil "github.com/ruslanSorokin/lock-manager/internal/pkg/util/grpc"
 )
 
 type RecoveryMetric struct {
