+++ internal/pkg/util/prom/wire.go
